refactoring edit dayid data type

+++ source/DomainModel/Entities/SettingsEntities/Weekday.cs
@@ -3,7 +3,7 @@
 namespace DomainModel.Entities.SettingsEntities;
 public class Weekday
 {
-    public short Id { get; set; }
+    public byte Id { get; set; }
 
     public virtual ICollection<WeekdaysTranslation> WeekdaysTranslations { get; set; } = new List<WeekdaysTranslation>();
 }
